refactor: reorganize UI components into common directory and update import paths
