+++ src/components/common/CourseCard.tsx
