+++ src/components/common/LearningPath.tsx
