+++ src/components/courses/Courses.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useQuery } from "@apollo/client";
 import { useNavigate } from "react-router-dom";
-import CourseCard from "../ui/CourseCard";
+import CourseCard from "../common/CourseCard";
 import { GET_COURSES_WITH_LEARNING_PATHS } from '../../graphql/queries/courses';
 import { useAuth } from '../../context/AuthContext';
